Moved everything pertaining to AD simulation into src folder
- Also added logic to write instructions into a json file

+++ ADsimulation/src/tests/test.cpp
@@ -164,8 +164,8 @@ int main() {
     gameState = GameState();
     gameState = run(gameState, [](GameState& st) {return st.canBuyNextDimboost();}, true);
     cout << endl << gameState << endl;
-    vector<int> instructions = gameState.getCompletedInstructions();
-    for (int instruction : instructions) {
+    vector<double> instructions = gameState.getCompletedInstructions();
+    for (double instruction : instructions) {
         cout << instruction << " ";
     }
     cout << endl;
